📐 Fmt settings & apply

+++ src/tkn/ERC20.sol
@@ -8,16 +8,17 @@ contract ERC20 {
     /// ERC20 events
     /// -----------------------------------------------------------------------
 
-    event Transfer(address indexed from, address indexed to, uint256 amt);
-    event Approval(address indexed from, address indexed to, uint256 amt);
+    event Transfer(address indexed from, address indexed to, uint amt);
+    event Approval(address indexed from, address indexed to, uint amt);
 
     /// -----------------------------------------------------------------------
     /// ERC20 storage
     /// -----------------------------------------------------------------------
 
-    mapping(address => mapping(address => uint256)) public allowance;
-    mapping(address => uint256) public balanceOf;
-    uint256 public totalSupply;
+    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint) public balanceOf;
+
+    uint public totalSupply;
 
     /// -----------------------------------------------------------------------
     /// ERC20 metadata immutables
@@ -25,7 +26,7 @@ contract ERC20 {
 
     bytes32 immutable _name;
     bytes32 immutable _symbol;
-    uint256 public constant decimals = 18;
+    uint public constant decimals = 18;
 
     function name() public view returns (string memory) {
         return string(abi.encodePacked(_name));
@@ -48,13 +49,13 @@ contract ERC20 {
     /// ERC20 logic
     /// -----------------------------------------------------------------------
 
-    function approve(address to, uint256 amt) public payable returns (bool) {
+    function approve(address to, uint amt) public payable returns (bool) {
         allowance[msg.sender][to] = amt;
         emit Approval(msg.sender, to, amt);
         return true;
     }
 
-    function transfer(address to, uint256 amt) public payable returns (bool) {
+    function transfer(address to, uint amt) public payable returns (bool) {
         balanceOf[msg.sender] -= amt;
         unchecked {
             balanceOf[to] += amt;
@@ -63,7 +64,7 @@ contract ERC20 {
         return true;
     }
 
-    function transferFrom(address from, address to, uint256 amt) public payable returns (bool) {
+    function transferFrom(address from, address to, uint amt) public payable returns (bool) {
         allowance[from][msg.sender] -= amt;
         balanceOf[from] -= amt;
         unchecked {
@@ -77,7 +78,7 @@ contract ERC20 {
     /// Internal mint/burn logic
     /// -----------------------------------------------------------------------
 
-    function _mint(address to, uint256 amt) internal {
+    function _mint(address to, uint amt) internal {
         totalSupply += amt;
         unchecked {
             balanceOf[to] += amt;
@@ -85,7 +86,7 @@ contract ERC20 {
         emit Transfer(address(0), to, amt);
     }
 
-    function _burn(address from, uint256 amt) internal {
+    function _burn(address from, uint amt) internal {
         balanceOf[from] -= amt;
         unchecked {
             totalSupply -= amt;
