+++ src/util/Refunded.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity 0.8.20;
 
-import {safeTransferETH} from "./SafeTransfer.sol";
-import "./ReentrancyGuard.sol";
+import {safeTransferETH} from './SafeTransfer.sol';
+import './ReentrancyGuard.sol';
 
 /// @dev Gas refund logic
 /// @author Zolidity
 abstract contract Refunded is ReentrancyGuard {
     error GasMax(address caller);
 
-    uint256 constant BASE_COST = 25433;
-    uint256 constant GAS_PRICE_MAX = 4e10;
+    uint constant BASE_COST = 25433;
+    uint constant GAS_PRICE_MAX = 4e10;
 
     receive() external payable virtual {}
 
@@ -18,7 +18,7 @@ abstract contract Refunded is ReentrancyGuard {
     ///      benefit most from refund
     modifier refunded() {
         // Memo starting gas
-        uint256 refund = gasleft();
+        uint refund = gasleft();
 
         setReentrancyGuard();
