+++ test/ERC20.t.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity 0.8.20;
 
-import "forge-std/Test.sol";
-import "./util/mock/MockERC20.sol";
+import 'forge-std/Test.sol';
+import './util/mock/MockERC20.sol';
 
 contract ERC20Test is Test {
     using stdStorage for StdStorage;
 
-    bytes32 constant META = keccak256("META");
-    uint256 constant ethSum = 100 ether;
+    bytes32 constant META = keccak256('META');
+    uint constant ethSum = 100 ether;
     address immutable alice = vm.addr(1);
 
     MockERC20 tkn;
@@ -16,7 +16,7 @@ contract ERC20Test is Test {
     constructor() payable {}
 
     function setUp() public payable {
-        console.log(unicode"🧪 Testing ERC20...");
+        console.log(unicode'🧪 Testing ERC20...');
         tkn = new MockERC20(META, META);
         tkn.mint(alice, ethSum);
     }
@@ -25,8 +25,8 @@ contract ERC20Test is Test {
         new MockERC20(META, META);
     }
 
-    function _safeTransfer(address from, address to, uint256 amt) internal {
-        uint256 balance = tkn.balanceOf(from);
+    function _safeTransfer(address from, address to, uint amt) internal {
+        uint balance = tkn.balanceOf(from);
         amt = balance < amt ? balance : amt;
 
         if (from != address(tkn)) {
@@ -36,8 +36,8 @@ contract ERC20Test is Test {
         tkn.transfer(to, amt);
     }
 
-    function testTransfer(address bob, uint256 amt) public payable {
-        uint256 alice_balance = tkn.balanceOf(alice);
+    function testTransfer(address bob, uint amt) public payable {
+        uint alice_balance = tkn.balanceOf(alice);
         if (amt > alice_balance) amt = alice_balance; // only transfer max possible amount
 
         vm.prank(alice);
@@ -49,8 +49,8 @@ contract ERC20Test is Test {
         }
     }
 
-    function testApprove(address bob, uint256 amt) public payable {
-        uint256 aliceBalance = tkn.balanceOf(alice);
+    function testApprove(address bob, uint amt) public payable {
+        uint aliceBalance = tkn.balanceOf(alice);
         amt = amt > aliceBalance ? aliceBalance : amt;
 
         vm.prank(alice);
@@ -59,8 +59,8 @@ contract ERC20Test is Test {
         assertEq(tkn.allowance(alice, bob), amt);
     }
 
-    function testTransferFrom(address bob, uint256 amt) public payable {
-        uint256 aliceBalance = tkn.balanceOf(alice);
+    function testTransferFrom(address bob, uint amt) public payable {
+        uint aliceBalance = tkn.balanceOf(alice);
         amt = amt > aliceBalance ? aliceBalance : amt;
 
         testApprove(bob, amt);
@@ -71,8 +71,8 @@ contract ERC20Test is Test {
         assertEq(tkn.balanceOf(bob), amt);
     }
 
-    function testMint(address bob, uint256 amt) public payable {
-        uint256 max_mint_value = type(uint256).max - ethSum;
+    function testMint(address bob, uint amt) public payable {
+        uint max_mint_value = type(uint).max - ethSum;
         amt = amt > max_mint_value ? max_mint_value : amt;
 
         tkn.mint(bob, amt);
@@ -81,8 +81,8 @@ contract ERC20Test is Test {
         assertEq(tkn.balanceOf(bob), amt);
     }
 
-    function testBurn(address bob, uint256 amt) public payable {
-        uint256 max_burn_value = tkn.balanceOf(bob);
+    function testBurn(address bob, uint amt) public payable {
+        uint max_burn_value = tkn.balanceOf(bob);
         amt = amt > max_burn_value ? max_burn_value : amt;
 
         vm.prank(bob);
